LayoutTests/crypto/webkitSubtle/aes-cbc-192-encrypt-decrypt-expected.txt

CONSOLE MESSAGE: line 22: WebKitSubtleCrypto is deprecated. Please use SubtleCrypto instead.
Test AES-CBC encrypt and decrypt functions with a 256 bit key.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Importing a raw AES key from string literal...
PASS key.type is 'secret'
PASS key.extractable is true
PASS key.algorithm.name is 'AES-CBC'
PASS key.algorithm.length is 192
PASS key.usages is ['decrypt', 'encrypt']
Using the key to encrypt plaintext...
PASS bytesToHexString(new Uint8Array(encryptionResult)) is '4f021db243bc633d7178183a9fa071e8b4d9ada9ad7dedf4e5e738763f69145a571b242012fb7ae07fa9baac3df102e008b0e27988598881d920a9e64f5615cd612ccd79224b350935d45dd6a98f8176'
Decrypting it back...
PASS new Uint8Array(decryptionResult) is plaintext
PASS successfullyParsed is true

TEST COMPLETE
